requirements.txt
streamlit==1.38.0
opencv-python-headless==4.10.0.84
numpy==1.26.4
scikit-image==0.24.0
matplotlib==3.9.2
pandas==2.2.2
pillow==10.4.0